# Optimize binary operations on bigints