--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,7 +15,7 @@ rustc_version = "0.2"
 rustc-hex = { version = "1.0", optional = true }
 heapsize = { version = "0.4", optional = true }
 byteorder = { version = "1", default-features = false }
-crunchy = { path = "../crunchy" }
+crunchy = "0.1.5"
 
 [features]
 heapsizeof = ["heapsize", "std"]